OpenXR fixes, some other bugs
Fix shadows for headset renderer
Fix builtin ImGui Windows (tools)
Fix OpenXR controller
Fix multiple selection with shift
Fix Ritter's bounding sphere
Fix node tree window drag on drop
- erhe.ini: added scene / object scale
- erhe now requires std::filesystem (gcc 8)
- erhe now required std::optional
Add Commands window
Improve sphere visualization

+++ src/erhe/application/rendergraph/rendergraph.cpp
@@ -266,7 +266,7 @@ void Rendergraph::automatic_layout()
     for (const auto& node : m_nodes)
     {
         const int depth = node->get_depth();
-        if (node_count_per_depth.size() < depth + 1)
+        if (node_count_per_depth.size() < static_cast<std::size_t>(depth) + 1)
         {
             node_count_per_depth.resize(depth + 1);
         }
